move all bot functions into pebbles.py
